+++ module_bet/src/main/java/arch/cayenne/module/bet/repo/BetResultRepository.kt
@@ -8,6 +8,6 @@ import kotlinx.coroutines.Dispatchers
 class BetResultRepository(private val betDao: BetDao): BaseRepository() {
     override val scope: CoroutineScope = CoroutineScope(Dispatchers.IO)
 
-    fun observeBetById(id: Int) = betDao.observeBetById(id)
+    fun observeBetById(id: Long) = betDao.observeBetById(id)
     fun observeComboBet() = betDao.observeComboBet()
 }
