+++ module_bet/src/main/java/arch/cayenne/module/bet/viewmodel/BetResultViewModel.kt
@@ -22,7 +22,7 @@ class BetResultViewModel(private val repo: BetResultRepository): BaseViewModel()
         _onBetModeListener.value = Pair(type, status)
     }
 
-    fun setBetSheet(id: Int?) {
+    fun setBetSheet(id: Long?) {
         viewModelScope.launch {
             if (id == null) {
                 launch {
